Implements Laser Beams

+++ src/main/java/model/Position.java
@@ -2,27 +2,27 @@ package model;
 
 public class Position {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Position(int x, int y){
+    public Position(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
